refactor: rename ad/semiring.rs to ad/elementwise_tier1.rs
The file holds mainline StdTensorOp AD rules for Add / Mul / Neg /
Conj, not rules for the non-mainline SemiringOp<Alg>. The previous
name collided with the deprecated SemiringOp module; the new name
reflects the actual semantics and matches the elementwise_tier*
naming pattern.
